// src/types/generated/JumpHost.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AuthConfig } from "./AuthConfig";
import type { SshEndpoint } from "./SshEndpoint";

/**
 * A bastion/jump host to hop through before reaching the SSH server.
 */
export type JumpHost = { endpoint: SshEndpoint, auth: AuthConfig, };